requirements.txt
Flask==2.2.3
flask_cors==3.0.10
numpy==1.24.2
pdf2image==1.16.3
Pillow==9.4.0
PyPDF2==3.0.1
pytesseract==0.3.10
Werkzeug==2.2.3
